# Conflicts:
#	.idea/misc.xml
#	app/src/main/AndroidManifest.xml

+++ app/src/main/java/qubictech/iot/automation/broilerfirm/data/model/Dht11Data.kt
@@ -1,10 +1,10 @@
 package qubictech.iot.automation.broilerfirm.data.model
 
 data class Dht11Data(
-    val heat_index: Double?,
+    val heat_index: Float?,
     val humidity: Int?,
-    val temp_c: Double?,
-    val temp_f: Double?,
+    val temp_c: Float?,
+    val temp_f: Float?,
     var timestamp: Long?
 ) {
     constructor() : this(null, null, null, null, null)
